requirements.txt
amqp==5.0.7
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
blessings==1.7
bpython==0.22.1
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
curtsies==0.3.10
cwcwidth==0.1.5
Django==3.2.10
django-celery-beat==2.2.1
django-filter==21.1
django-restframework==0.0.1
django-tenants==3.4.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.9
djangorestframework==3.13.0
greenlet==1.1.2
idna==3.3
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
Pygments==2.10.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.27
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
